Fix WasiCtxBuilder so that stdin, stdout, and stderr accept files.
A regression in #1561 caused "invalid argument" errors when `stdin`, `stdout`,
and `stderr` are given a file handle.

This was missed in Wasmtime CI because previously only a pipe handle was being
used.

It was caught in the .NET implementation CI because it uses file handles for
its WASI tests.

Fixes #1735.